testing if verification is done before sending sign up email
