+++ server/src/subscribers/newUserSubscription.js
@@ -2,27 +2,13 @@ const Queue = require('bull');
 const EmailServices = require('../services/EmailServices');
 const jwt = require('jsonwebtoken');
 
-const newUserSubscription = new Queue('newUser', {
+const newUserSubscription = new Queue('existingUser', {
   redis: { port: process.env.REDIS_PORT, host: 'redis', password: process.env.REDIS_PASSWORD },
 });
 
 newUserSubscription.process((job, done) => {
   try {
-    // sending user email verification link
-    const emailToken = jwt.sign(
-      job.data.email,
-      process.env.JWT_RESET_TOKEN,
-    );
-    const url = process.env.CLIENT_BASE_URL + '/verify-user-email/' + job.data.email + '/' + emailToken;
-    EmailServices.sendSimpleEmail(
-       [job.data.email],
-       '',
-       'Please use this URL to confirm your email: ' + url,
-       'F!rosh Email Confirmation',
-       'tech@orientation.skule.ca',
-    );
-
-    // sending user creation email
+    // sending successful user creation email
     const result = EmailServices.sendTemplateEmail(
       {},
       'signup_confirmation',
